pin dependency `awesome_print` to v1.7.0 to avoid bug in v1.8.0
The `awesome_print` release v1.8.0 contains a bug that causes the Logstash
process to crash while loading if an exception is raised while the library
attempts to load its user-defined defaults from disk (this can occur if the
`HOME` environment variable is unset, or if the JVM fails to read the file
at `${HOME}/.aprc`).

While the bug has been [fixed][] on the `awesome_print` master branch, until
a release is rolled, pin our dependency to an earlier version to avoid the
bug.

[fixed]: https://github.com/awesome-print/awesome_print/commit/62033856f8ff083